Upgrade to .NET 10 on Iteration 2

+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/Services/ImageAzureStorage.cs
@@ -1,35 +1,35 @@
-﻿using eShopModernizedMVC.Models;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
+using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
+using eShopModernizedMVC.Models;
+using Microsoft.AspNetCore.Http;
 using System;
 using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
-using System.Web;
 
 namespace eShopModernizedMVC.Services
 {
     public class ImageAzureStorage : IImageService
     {
-
-        private readonly CloudStorageAccount _storageAccount;
+        private readonly BlobServiceClient _blobServiceClient;
+        private readonly string _containerName = "pics";
 
         public ImageAzureStorage()
         {
-            _storageAccount = CloudStorageAccount.Parse(CatalogConfiguration.StorageConnectionString);
+            _blobServiceClient = new BlobServiceClient(CatalogConfiguration.StorageConnectionString);
         }
 
         public string BaseUrl()
         {
-            return _storageAccount.BlobStorageUri.PrimaryUri.ToString();
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            return containerClient.Uri.ToString() + "/";
         }
 
         public string BuildUrlImage(CatalogItem item)
         {
             if (string.IsNullOrEmpty(item.PictureFileName))
                 return UrlDefaultImage();
-
-            return _storageAccount.BlobStorageUri.PrimaryUri + "pics/" + item.Id + "/" + item.PictureFileName;
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            return containerClient.Uri + "/" + item.Id + "/" + item.PictureFileName;
         }
 
         public void Dispose()
@@ -38,84 +38,53 @@ namespace eShopModernizedMVC.Services
 
         public void InitializeCatalogImages()
         {
-            CloudBlobClient blobClient = _storageAccount.CreateCloudBlobClient();
-            CloudBlobContainer container = blobClient.GetContainerReference("pics");
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            containerClient.CreateIfNotExists(PublicAccessType.Blob);
 
-            container.CreateIfNotExists();
-
-            BlobContainerPermissions permissions = container.GetPermissions();
-            permissions.PublicAccess = BlobContainerPublicAccessType.Blob;
-            container.SetPermissions(permissions);
-
-
-            Parallel.ForEach(container.ListBlobs().Where(x => x is CloudBlob), x => ((CloudBlob)x).Delete());
-            var webRoot = HttpContext.Current.Server.MapPath("~/Pics");
-
-            for (int i = 1; i <= 12; i++)
+            // Delete existing blobs
+            foreach (var blob in containerClient.GetBlobs())
             {
-                var path = Path.Combine(webRoot, i + ".png");
-                var blobName = i + "/" + i + ".png";
-                UpLoadImageFromFile(container, blobName, path, "image/png");
-
+                containerClient.GetBlobClient(blob.Name).Delete();
             }
-            var defaultImagePath = Path.Combine(webRoot, "default.png");
-            UpLoadImageFromFile(container, "temp/default.png", defaultImagePath, "image/png");
-
-
         }
 
         public void UpdateImage(CatalogItem item)
         {
-            CloudBlobClient blobClient = _storageAccount.CreateCloudBlobClient();
-            CloudBlobContainer container = blobClient.GetContainerReference("pics");
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            var folder = item.TempImageName?.Replace("/pics/", string.Empty) ?? string.Empty;
+            var tempBlob = containerClient.GetBlobClient(folder);
 
-            var folder = item.TempImageName.Replace("/pics/", string.Empty);
-
-            CloudBlockBlob tempBlob = container.GetBlockBlobReference(folder);
-
-            var blockBlobs = container.ListBlobs(prefix: item.Id + "/").OfType<CloudBlockBlob>();
-            foreach (var blockBlob in blockBlobs)
+            // Delete existing blobs for item
+            foreach (var blob in containerClient.GetBlobs(prefix: item.Id + "/"))
             {
-                blockBlob.Delete();
+                containerClient.GetBlobClient(blob.Name).Delete();
             }
 
-            var fileName = Path.GetFileName(item.TempImageName);
-            CloudBlockBlob imageBlob = container.GetBlockBlobReference(item.Id + "/" + fileName);
-
-            imageBlob.StartCopy(tempBlob);
-            tempBlob.Delete();
+            if (!string.IsNullOrEmpty(folder))
+            {
+                var fileName = Path.GetFileName(item.TempImageName);
+                var targetBlob = containerClient.GetBlobClient(item.Id + "/" + fileName);
+                targetBlob.StartCopyFromUri(tempBlob.Uri);
+                tempBlob.Delete();
+            }
         }
 
-        public string UploadTempImage(HttpPostedFile file, int? catalogItemId)
+        public async Task<string> UploadTempImageAsync(IFormFile file, int? catalogItemId)
         {
             string path = catalogItemId.HasValue ? catalogItemId + "/temp/" : "temp/" + Guid.NewGuid().ToString() + "/";
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            var blobClient = containerClient.GetBlobClient(path + file.FileName.ToLower());
 
-            CloudBlobClient blobClient = _storageAccount.CreateCloudBlobClient();
-            CloudBlobContainer container = blobClient.GetContainerReference("pics");
-            CloudBlockBlob blockBlob = container.GetBlockBlobReference(path + file.FileName.ToLower());
+            using var stream = file.OpenReadStream();
+            await blobClient.UploadAsync(stream, new BlobHttpHeaders { ContentType = file.ContentType });
 
-            blockBlob.Properties.ContentType = file.ContentType;
-            file.InputStream.Seek(0, SeekOrigin.Begin);
-            blockBlob.UploadFromStream(file.InputStream);
-
-            return blockBlob.Uri.ToString();
+            return blobClient.Uri.ToString();
         }
 
         public string UrlDefaultImage()
         {
-            return _storageAccount.BlobStorageUri.PrimaryUri + "pics/temp/default.png";
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            return containerClient.Uri + "/temp/default.png";
         }
-
-        private void UpLoadImageFromFile(CloudBlobContainer container, string blobName, string filePath, string contentType)
-        {
-            var fileStream = File.OpenRead(filePath);
-            fileStream.Seek(0, SeekOrigin.Begin);
-
-            CloudBlockBlob blockBlob = container.GetBlockBlobReference(blobName);
-            blockBlob.Properties.ContentType = contentType;
-            blockBlob.UploadFromStream(fileStream);
-        }
-
-
     }
 }
